Started moving implementation from Interaction.h to Interaction.cc.
